requirements.txt
async-lru==1.0.3
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cacheable-iterators==0.1.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
graphviz==0.20
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.4
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydot==1.4.2
pyparsing==3.0.9
pytest==7.1.2
requests==2.28.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
